+++ NanoXLSX/Styles/Font.cs
@@ -24,11 +24,6 @@ namespace NanoXLSX.Styles
     {
         #region constants
         /// <summary>
-        /// Default font family as constant
-        /// </summary>
-        public static readonly string DEFAULT_FONT_NAME = "Calibri";
-
-        /// <summary>
         /// Maximum possible font size
         /// </summary>
         public static readonly float MIN_FONT_SIZE = 1f;
@@ -42,7 +37,20 @@ namespace NanoXLSX.Styles
         /// Default font size
         /// </summary>
         public static readonly float DEFAULT_FONT_SIZE = 11f;
-        #endregion
+        
+        /// <summary>
+        /// The default font name that is declared as Major Font (See <see cref="Font.SchemeValue"/>)
+        /// </summary>
+        public static readonly string DEFAULT_MAJOR_FONT = "Calibri Light";
+        /// <summary>
+        /// The default font name that is declared as Minor Font (See <see cref="Font.SchemeValue"/>)
+        /// </summary>
+        public static readonly string DEFAULT_MINOR_FONT = "Calibri";
+
+        /// <summary>
+        /// Default font family as constant
+        /// </summary>
+        public static readonly string DEFAULT_FONT_NAME = DEFAULT_MINOR_FONT;
 
         /// <summary>
         /// Default font family
@@ -58,6 +66,7 @@ namespace NanoXLSX.Styles
         /// Default vertical alignment
         /// </summary>
         public static readonly VerticalTextAlignValue DEFAULT_VERTICAL_ALIGN = VerticalTextAlignValue.none;
+        #endregion
 
         #region enums
 
@@ -167,11 +176,8 @@ namespace NanoXLSX.Styles
             get { return name; }
             set
             {
-                if (string.IsNullOrEmpty(value) && !StyleRepository.Instance.ImportInProgress)
-                {
-                    throw new StyleException("The font name was null or empty");
-                }
                 name = value;
+                ValidateFontScheme();
             }
         }
         /// <summary>
@@ -220,7 +226,31 @@ namespace NanoXLSX.Styles
         }
         #endregion
 
-        #region methods            
+        #region methods  
+
+        /// <summary>
+        /// Validates the font name and sets the scheme automatically
+        /// </summary>
+        private void ValidateFontScheme()
+        {
+            if ((string.IsNullOrEmpty(name)) && !StyleRepository.Instance.ImportInProgress)
+            {
+                throw new StyleException("The font name was null or empty");
+            }
+            if (name.Equals(DEFAULT_MINOR_FONT))
+            {
+                Scheme = SchemeValue.minor;
+            }
+            else if (name.Equals(DEFAULT_MAJOR_FONT))
+            {
+                Scheme = SchemeValue.major;
+            }
+            else
+            {
+                Scheme = SchemeValue.none;
+            }
+        }
+
         /// <summary>
         /// Override toString method
         /// </summary>