+++ NanoXLSX/Worksheet.cs
@@ -1289,26 +1289,26 @@ namespace NanoXLSX
                     return cells.Max(x => x.Value.ColumnNumber);
                 }
             }
-            List<Cell> nonEmptyCells = cells.Values.Where(x => x.Value != null).ToList();
+            List<Cell> nonEmptyCells = cells.Values.Where(x => x.Value != null && x.Value.ToString() != string.Empty).ToList();
             if (nonEmptyCells.Count == 0)
             {
                 return -1;
             }
             if (row && min)
             {
-                return nonEmptyCells.Where(x => x.Value.ToString() != string.Empty).Min(x => x.RowNumber);
+                return nonEmptyCells.Min(x => x.RowNumber);
             }
             else if (row)
             {
-                return nonEmptyCells.Where(x => x.Value.ToString() != string.Empty).Max(x => x.RowNumber);
+                return nonEmptyCells.Max(x => x.RowNumber);
             }
             else if (min)
             {
-                return nonEmptyCells.Where(x => x.Value.ToString() != string.Empty).Min(x => x.ColumnNumber);
+                return nonEmptyCells.Min(x => x.ColumnNumber);
             }
             else
             {
-                return nonEmptyCells.Where(x => x.Value.ToString() != string.Empty).Max(x => x.ColumnNumber);
+                return nonEmptyCells.Max(x => x.ColumnNumber);
             }
         }
 
@@ -2009,6 +2009,41 @@ namespace NanoXLSX
                 Column c = new Column(columnNumber);
                 c.Width = width;
                 columns.Add(columnNumber, c);
+            }
+        }
+
+        /// <summary>
+        /// Sets the default column style of the passed column address
+        /// </summary>
+        /// <param name="columnAddress">Column address (A - XFD)</param>
+        /// <param name="style">Style to set as default. If null, the style is cleared</param>
+        /// <returns>Assigned style or null if cleared</returns>
+        /// <exception cref="RangeException">Throws a RangeException:<br></br>a) If the passed column address is out of range<br></br>b) if the column width is out of range (0 - 255.0)</exception>
+        public Style SetColumnDefaultStyle(string columnAddress, Style style)
+        {
+            int columnNumber = Cell.ResolveColumn(columnAddress);
+            return SetColumnDefaultStyle(columnNumber, style);
+        }
+        /// <summary>
+        /// Sets the default column style of the passed column number (zero-based)
+        /// </summary>
+        /// <param name="columnNumber">Column number (zero-based, from 0 to 16383)</param>
+        /// <param name="style">Style to set as default. If null, the style is cleared</param>
+        /// <returns>Assigned style or null if cleared</returns>
+        /// <exception cref="RangeException">Throws a RangeException:<br></br>a) If the passed column number is out of range<br></br>b) if the column width is out of range (0 - 255.0)</exception>
+        public Style SetColumnDefaultStyle(int columnNumber, Style style)
+        {
+            Cell.ValidateColumnNumber(columnNumber);
+            if (this.columns.ContainsKey(columnNumber))
+            {
+                return this.columns[columnNumber].SetDefaultColumnStyle(style);
+            }
+            else
+            {
+                Column c = new Column(columnNumber);
+                Style returnStyle = c.SetDefaultColumnStyle(style);
+                this.columns.Add(columnNumber, c);
+                return returnStyle;
             }
         }